moved test to separate test file
