+++ cypress/e2e/loans/loan-dates.cy.js
@@ -16,12 +16,6 @@ import MultipieceCheckOut from '../../support/fragments/checkout/modals/multipie
 import ServicePoints from '../../support/fragments/settings/tenant/servicePoints/servicePoints';
 import InventoryInstance from '../../support/fragments/inventory/inventoryInstance';
 import InventoryInstances from '../../support/fragments/inventory/inventoryInstances';
-import Checkout from '../../support/fragments/checkout/checkout';
-import { Locations } from '../../support/fragments/settings/tenant/location-setup';
-import Location from '../../support/fragments/settings/tenant/locations/newLocation';
-import SearchPane from '../../support/fragments/circulation-log/searchPane';
-import { MultiColumnListCell } from '../../../interactors';
-import SearchResults from '../../support/fragments/circulation-log/searchResults';
 
 const folioInstances = InventoryInstances.generateFolioInstances({
   properties: { missingPieces: '3', numberOfMissingPieces: '3' },
@@ -150,101 +144,6 @@ describe('loan dates', () => {
       LoansPage.openChangeDueDateForm();
       ChangeDueDateForm.verifyRequestsCount('1');
       ChangeDueDateForm.verifyWarning('Item has been recalled');
-    },
-  );
-});
-
-describe('loan date and time', () => {
-  const testData = {
-    folioInstances: InventoryInstances.generateFolioInstances(),
-    servicePoint: ServicePoints.getDefaultServicePointWithPickUpLocation(),
-    requestsId: '',
-  };
-  let userData;
-  let ITEM_BARCODE;
-
-  before('create inventory instance', () => {
-    cy.getAdminToken();
-    ServicePoints.createViaApi(testData.servicePoint);
-    testData.defaultLocation = Location.getDefaultLocation(testData.servicePoint.id);
-    Location.createViaApi(testData.defaultLocation).then((location) => {
-      InventoryInstances.createFolioInstancesViaApi({
-        folioInstances: testData.folioInstances,
-        location,
-      });
-    });
-    cy.createTempUser([
-      Permissions.circulationLogAll.gui,
-      Permissions.uiInventoryViewInstances.gui,
-      Permissions.uiUsersViewLoans.gui,
-      Permissions.inventoryAll.gui,
-    ])
-      .then((userProperties) => {
-        userData = userProperties;
-      })
-      .then(() => {
-        ITEM_BARCODE = testData.folioInstances[0].barcodes[0];
-        UserEdit.addServicePointViaApi(testData.servicePoint.id, userData.userId);
-
-        Checkout.checkoutItemViaApi({
-          itemBarcode: ITEM_BARCODE,
-          servicePointId: testData.servicePoint.id,
-          userBarcode: userData.barcode,
-        });
-        cy.login(userData.username, userData.password, {
-          path: TopMenu.circulationLogPath,
-          waiter: SearchPane.waitLoading,
-        });
-      });
-  });
-
-  after('Delete all data', () => {
-    CheckinActions.checkinItemViaApi({
-      itemBarcode: ITEM_BARCODE,
-      servicePointId: testData.servicePoint.id,
-    });
-    UserEdit.changeServicePointPreferenceViaApi(userData.userId, [testData.servicePoint.id]);
-    ServicePoints.deleteViaApi(testData.servicePoint.id);
-    InventoryInstances.deleteInstanceViaApi({
-      instance: testData.folioInstances[0],
-      servicePoint: testData.servicePoint,
-      shouldCheckIn: true,
-    });
-    Locations.deleteViaApi(testData.defaultLocation);
-    Users.deleteViaApi(userData.userId);
-  });
-
-  const navigateToCircLogAndSearchItem = (barcode) => {
-    // Select the Circulation log app
-    cy.visit(TopMenu.circulationLogPath);
-    // "Loan" accordion => Expand "Loan" accordion and Select "Checked Out" and "Checked In" => Search for an item
-    SearchPane.searchByCheckedOut();
-    SearchPane.checkResultSearch({
-      itemBarcode: barcode,
-      circAction: 'Checked out',
-    });
-    SearchPane.searchByItemBarcode(barcode);
-  };
-
-  it(
-    'C350710 Check date and time -- loans (firebird) (TaaS)',
-    { tags: [TestTypes.extendedPath, DevTeams.firebird] },
-    () => {
-      navigateToCircLogAndSearchItem(ITEM_BARCODE);
-      SearchPane.verifyResultCells(true);
-      cy.wrap(MultiColumnListCell({ row: 0, columnIndex: 4 }).text()).as('date');
-      // Click to "Item barcode" on the row with Circ Action "Checked out" => View "Checked Out" date and time displayed on Items view page
-      SearchPane.goToItemDetails(ITEM_BARCODE);
-      cy.get('@date').then((date) => {
-        InventoryInstance.verifyCheckedOutDate(date);
-        // Go to the result list with items again
-        navigateToCircLogAndSearchItem(ITEM_BARCODE);
-        // Click on ellipsed Action dropdown from the Action column => click on 'Loan Details'
-        SearchResults.chooseActionByRow(0, 'Loan details');
-        LoansPage.waitLoading();
-        // Check the date and time displayed for 'Checked Out' of loan
-        LoansPage.checkLoanDate(date);
-      });
     },
   );
 });
